requirements.txt
alpaca-trade-api
streamlit
pandas
numpy
python-dotenv
plotly


